Whitelist some compiler options for module sub-compile forwarding
Summary:

// This file defines the language options. Users of this file must
// define the LANGOPT macro to make use of this information.
//
// Optionally, the user may also define:
//
// BENIGN_LANGOPT: for options that don't affect the construction of the AST in
//     any way (that is, the value can be different between an implicit module
//     and the user of that module).
//
// COMPATIBLE_LANGOPT: for options that affect the construction of the AST in
//     a way that doesn't prevent interoperability (that is, the value can be
//     different between an explicit module and the user of that module).

Clang compiles modules in a separate internal compiler instance, with its
own args. resetNonModularOptions() removes benign langopts from args sent
to this instance, but we want the ability to forward these.

I think the actual definition of benign doesn't really fit here. For
example, I think the real reason --fmodules-codegen was benign was that
the guy who wrote thought it wasn't practical for builds based on implicit
modules.

We don't use all of these, but I do want the ability to experiment with
them.